refactor: move files in core package
