+++ src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/core/AggregateExpression.kt
@@ -1,4 +1,4 @@
-package io.github.filippovissani.kotlin_distributed_dfrp
+package io.github.filippovissani.kotlin_distributed_dfrp.core
 
 import kotlinx.coroutines.flow.*
 
@@ -32,7 +32,7 @@ fun <T, R> AggregateExpression<T>.map(transform: (T) -> R): AggregateExpression<
 
 fun <T1, T2, R> combine(exp1: AggregateExpression<T1>, exp2: AggregateExpression<T2>, transform: (T1, T2) -> R): AggregateExpression<R> {
     return AggregateExpression.of { context, path ->
-        combine(exp1.compute(path, context), exp2.compute(path, context)){ result1, result2 ->
+        combine(exp1.compute(path, context), exp2.compute(path, context)) { result1, result2 ->
             ExportTree(transform(result1.root, result2.root))
         }
     }
